Gradle 5.0 and other updates JUnit 4 extraction compiles
