+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,5 @@
-#Fri Aug 04 11:59:26 CEST 2017
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.0-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionSha256Sum=b49c6da1b2cb67a0caf6c7480630b51c70a11ca2016ff2f555eaeda863143a29
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.2-bin.zip
