+++ spock-core/core.gradle
@@ -12,7 +12,7 @@ configurations { coreConsoleRuntime }
 
 dependencies {
   compile libs.groovy // easiest way to add Groovy dependency to POM
-  compile libs.junit
+  compile libs.junit4
 
   compile libs.jetbrainsAnnotations, optional
   compile libs.ant, optional
