+++ spock-core/src/main/resources/META-INF/services/org.spockframework.runtime.extension.IGlobalExtension
@@ -1,5 +1,3 @@
 org.spockframework.runtime.extension.builtin.IncludeExcludeExtension
-org.spockframework.runtime.extension.builtin.RuleExtension
-org.spockframework.runtime.extension.builtin.ClassRuleExtension
 org.spockframework.runtime.extension.builtin.OptimizeRunOrderExtension
 org.spockframework.report.log.ReportLogExtension
