+++ spock-junit4/src/main/resources/META-INF/services/org.spockframework.runtime.extension.IGlobalExtension
@@ -1,0 +1,2 @@
+org.spockframework.junit4.RuleExtension
+org.spockframework.junit4.ClassRuleExtension
