+++ spock-junit4/src/test/groovy/org/spockframework/junit4/junit/RulesAndInheritance.groovy
@@ -12,12 +12,14 @@
  * limitations under the License.
  */
 
-package org.spockframework.smoke.junit
+package org.spockframework.junit4.junit
 
 import org.junit.Rule
 import org.junit.rules.TestName
 
-import spock.lang.*
+import spock.lang.Issue
+import spock.lang.Shared
+import spock.lang.Specification
 
 @Issue("http://issues.spockframework.org/detail?id=98")
 class RulesAndInheritance extends RulesAndInheritanceBase {
